requirements.txt
asgiref==3.2.7
Django==3.0.6
django-taggit==1.3.0
isbnlib==3.10.1
pytz==2020.1
sqlparse==0.3.1
